+++ messaginginapp/src/androidTest/java/io/customer/messaginginapp/ModuleMessagingInAppTest.kt
@@ -2,12 +2,19 @@ package io.customer.messaginginapp
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import io.customer.commontest.BaseIntegrationTest
+import io.customer.messaginginapp.di.inAppMessaging
 import io.customer.messaginginapp.provider.InAppMessagesProvider
 import io.customer.messaginginapp.type.InAppEventListener
+import io.customer.sdk.CustomerIO
 import io.customer.sdk.CustomerIOConfig
+import io.customer.sdk.data.model.Region
+import io.customer.sdk.extensions.random
+import io.customer.sdk.hooks.HookModule
 import io.customer.sdk.hooks.HooksManager
 import io.customer.sdk.module.CustomerIOModule
 import io.customer.sdk.repository.preference.SitePreferenceRepository
+import java.lang.reflect.Field
+import org.amshove.kluent.shouldBe
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -45,6 +52,27 @@ internal class ModuleMessagingInAppTest : BaseIntegrationTest() {
     }
 
     @Test
+    fun initialize_givenComponentInitialize_expectGistToInitializeWithCorrectValuesAndHooks() {
+        module.initialize()
+
+        // verify gist is initialized
+        verify(gistInAppMessagesProvider).initProvider(
+            any(),
+            eq(cioConfig.siteId),
+            eq(cioConfig.region.code)
+        )
+
+        // verify hook was added
+        verify(hooksManager).add(eq(HookModule.MessagingInApp), any())
+
+        // verify events
+        verify(gistInAppMessagesProvider).subscribeToEvents(any(), any(), any())
+
+        // verify given event listener gets registered
+        verify(gistInAppMessagesProvider).setListener(eventListenerMock)
+    }
+
+    @Test
     fun initialize_givenProfilePreviouslyIdentified_expectGistToSetUserToken() {
         prefRepository.saveIdentifier("identifier")
 
@@ -59,5 +87,60 @@ internal class ModuleMessagingInAppTest : BaseIntegrationTest() {
 
         // verify gist sets userToken
         verify(gistInAppMessagesProvider).setUserToken(eq("identifier"))
+    }
+
+    @Test
+    fun initialize_givenNoProfileIdentified_expectGistNoUserSet() {
+        module.initialize()
+
+        // verify gist is initialized
+        verify(gistInAppMessagesProvider).initProvider(
+            any(),
+            eq(cioConfig.siteId),
+            eq(cioConfig.region.code)
+        )
+
+        // verify gist doesn't userToken
+        verify(gistInAppMessagesProvider, never()).setUserToken(any())
+    }
+
+    @Test
+    fun initialize_givenComponentInitializedWithOrganizationId_expectOrganizationIdToBeIgnored() {
+        // since `organizationId` is a private member, to check if its being used we have to use reflection
+        // this test will be deleted when the deprecated variable is removed
+
+        val orgId = String.random
+        val module = ModuleMessagingInApp(
+            organizationId = orgId
+        )
+        val fields = ModuleMessagingInApp::class.java.declaredFields
+        var organizationId: Field? = null
+        for (field in fields) {
+            if (field.name == "organizationId") {
+                organizationId = field
+                break
+            }
+        }
+        organizationId?.isAccessible = true
+        (organizationId?.get(module)) shouldBe null
+    }
+
+    @Test
+    fun whenDismissMessageCalledOnCustomerIO_thenDismissMessageIsCalledOnGist() {
+        // initialize the SDK
+        val customerIO = CustomerIO.Builder(
+            siteId = siteId,
+            apiKey = String.random,
+            region = Region.US,
+            appContext = application
+        ).apply {
+            overrideDiGraph = di
+        }.build()
+
+        // call dismissMessage on the CustomerIO instance
+        customerIO.inAppMessaging().dismissMessage()
+
+        // verify that the module's dismissMessage method was called
+        verify(gistInAppMessagesProvider).dismissMessage()
     }
 }
