--- messaginginapp/src/test/java/io/customer/messaginginapp/ModuleMessagingInAppTest.kt
@@ -1,130 +1,0 @@
-package io.customer.messaginginapp
-
-import io.customer.commontest.BaseUnitTest
-import io.customer.messaginginapp.di.inAppMessaging
-import io.customer.messaginginapp.provider.InAppMessagesProvider
-import io.customer.messaginginapp.type.InAppEventListener
-import io.customer.sdk.CustomerIO
-import io.customer.sdk.CustomerIOConfig
-import io.customer.sdk.data.model.Region
-import io.customer.sdk.extensions.random
-import io.customer.sdk.hooks.HookModule
-import io.customer.sdk.hooks.HooksManager
-import io.customer.sdk.module.CustomerIOModule
-import io.customer.sdk.repository.preference.SitePreferenceRepository
-import java.lang.reflect.Field
-import org.amshove.kluent.shouldBe
-import org.junit.Before
-import org.junit.Test
-import org.mockito.kotlin.any
-import org.mockito.kotlin.eq
-import org.mockito.kotlin.mock
-import org.mockito.kotlin.never
-import org.mockito.kotlin.verify
-
-internal class ModuleMessagingInAppTest : BaseUnitTest() {
-
-    private lateinit var module: ModuleMessagingInApp
-    private val gistInAppMessagesProvider: InAppMessagesProvider = mock()
-    private val hooksManager: HooksManager = mock()
-    private val eventListenerMock: InAppEventListener = mock()
-    private val prefRepository: SitePreferenceRepository
-        get() = di.sitePreferenceRepository
-
-    private val modules = hashMapOf<String, CustomerIOModule<*>>()
-
-    override fun setupConfig(): CustomerIOConfig = createConfig(
-        modules = modules
-    )
-
-    @Before
-    override fun setup() {
-        super.setup()
-
-        di.overrideDependency(InAppMessagesProvider::class.java, gistInAppMessagesProvider)
-        di.overrideDependency(HooksManager::class.java, hooksManager)
-
-        module = ModuleMessagingInApp(
-            moduleConfig = MessagingInAppModuleConfig.Builder().setEventListener(eventListenerMock)
-                .build(),
-            overrideDiGraph = di
-        )
-        modules[ModuleMessagingInApp.moduleName] = module
-    }
-
-    @Test
-    fun initialize_givenComponentInitialize_expectGistToInitializeWithCorrectValuesAndHooks() {
-        module.initialize()
-
-        // verify gist is initialized
-        verify(gistInAppMessagesProvider).initProvider(
-            any(),
-            eq(cioConfig.siteId),
-            eq(cioConfig.region.code)
-        )
-
-        // verify hook was added
-        verify(hooksManager).add(eq(HookModule.MessagingInApp), any())
-
-        // verify events
-        verify(gistInAppMessagesProvider).subscribeToEvents(any(), any(), any())
-
-        // verify given event listener gets registered
-        verify(gistInAppMessagesProvider).setListener(eventListenerMock)
-    }
-
-    @Test
-    fun initialize_givenNoProfileIdentified_expectGistNoUserSet() {
-        module.initialize()
-
-        // verify gist is initialized
-        verify(gistInAppMessagesProvider).initProvider(
-            any(),
-            eq(cioConfig.siteId),
-            eq(cioConfig.region.code)
-        )
-
-        // verify gist doesn't userToken
-        verify(gistInAppMessagesProvider, never()).setUserToken(any())
-    }
-
-    @Test
-    fun initialize_givenComponentInitializedWithOrganizationId_expectOrganizationIdToBeIgnored() {
-        // since `organizationId` is a private member, to check if its being used we have to use reflection
-        // this test will be deleted when the deprecated variable is removed
-
-        val orgId = String.random
-        val module = ModuleMessagingInApp(
-            organizationId = orgId
-        )
-        val fields = ModuleMessagingInApp::class.java.declaredFields
-        var organizationId: Field? = null
-        for (field in fields) {
-            if (field.name == "organizationId") {
-                organizationId = field
-                break
-            }
-        }
-        organizationId?.isAccessible = true
-        (organizationId?.get(module)) shouldBe null
-    }
-
-    @Test
-    fun whenDismissMessageCalledOnCustomerIO_thenDismissMessageIsCalledOnGist() {
-        // initialize the SDK
-        val customerIO = CustomerIO.Builder(
-            siteId = siteId,
-            apiKey = String.random,
-            region = Region.US,
-            appContext = application
-        ).apply {
-            overrideDiGraph = di
-        }.build()
-
-        // call dismissMessage on the CustomerIO instance
-        customerIO.inAppMessaging().dismissMessage()
-
-        // verify that the module's dismissMessage method was called
-        verify(gistInAppMessagesProvider).dismissMessage()
-    }
-}
